一般ユーザー：AttendanceListControllerを削除し、AttendanceControllerへ統合
